Changed a lot of things with Carousel
Filter is not needed, except we can just request each of the ids in the table and display them (every 5 seconds). There is a bug, where if there are multiple sections, then it displays the professor of the class multiple times. It can be fixed but this is the initial setup.